[Tech] Reviewed backend and fixed bugs
