+++ src/backend/storeManagers/steam/electronStores.ts
@@ -1,6 +1,7 @@
 import { TypeCheckedStoreBackend } from '../../electron_store'
 import CacheStore from '../../cache'
 import { ExtraInfo, GameInfo } from 'common/types'
+import { SteamInstallInfo } from 'common/types/steam'
 
 const installedGamesStore = new TypeCheckedStoreBackend(
   'steamInstalledGamesStore',
@@ -16,12 +17,15 @@ const configStore = new TypeCheckedStoreBackend('steamConfigStore', {
 
 const libraryStore = new CacheStore<GameInfo[], 'games'>('steam_library', null)
 
-// Caches the Steam storefront "appdetails" response (description, requirements,
-// release date, etc.) per app id so we don't hit the API on every page open.
-// The `_v2` suffix is a cache-schema bump: older entries predate the `background`
-// (splash) and `score` fields, so orphaning them forces a fresh fetch the next
-// time a game is viewed - making the new info show up right away instead of
-// after the cache's normal lifespan.
+// Caches the Steam storefront "appdetails" response
 const extraInfoStore = new CacheStore<ExtraInfo>('steam_extra_info_v2')
 
-export { configStore, installedGamesStore, libraryStore, extraInfoStore }
+const installInfoStore = new CacheStore<SteamInstallInfo>('steam_install_info')
+
+export {
+  configStore,
+  installedGamesStore,
+  libraryStore,
+  extraInfoStore,
+  installInfoStore
+}
